Add files via upload
unoAllTrial3 represents combining all the functionality of the project from the colors to the heart rate monitor. This currently prints out HR values and light color synchronously but it needs to be done asynchronously. 

nodemcutestApril6 represents the receiver. It is supposed to receive float variable as strings and insert them into a PHP script. 

nodemcutestMarch26TRIAL is the last working version of the Node MCU receiver code. 

unotestMarch26 TRIAL is the last working version of the Arduino Uno sender code.